+++ tests/end-to-end/utils/helpers.js
@@ -21,10 +21,13 @@ const runCacheAction = async (actionType = "refresh") => {
     await expect(page).toMatchTitle(`Cache action: ${actionType}`);
 };
 
-const runGeolocationTest = async (ip, saveInSession) => {
+const runGeolocationTest = async (ip, saveResult, brokenDb = false) => {
     let url = `/my-own-modules/crowdsec-php-lib/scripts/public/geolocation-test.php?ip=${ip}`;
-    if (saveInSession) {
-        url += `&session-save=1`;
+    if (saveResult) {
+        url += "&save-result=1";
+    }
+    if (brokenDb) {
+        url += "&broken-db=1";
     }
     await goToPublicPage(`${url}`);
     await page.waitForLoadState("networkidle");